Add files via upload
Uploaded the final version of the project for college capstone completion and IACIS presentation. I was doing some additional work with competitive environments, hence the CompMazeEnv.py file. Unfortunately, I was not able to finish this portion in time and will return to it later. However, the base simulation can still run via the simulation.py file.